fix(ci): bound fuzz memory/time so the fuzz job stops OOM-thrashing the runner
The fuzz job (the last red one) was cancelled mid-fuzz_json: on the no-swap
16GB hosted runner, 20000 mutations x 6 ASAN targets starved the runner agent
(the documented historical failure mode). All 6 targets pass locally with no
crash — it's a runner-resource issue, not a fuzz finding. Four mitigations:

- FUZZ_ASAN_ENV: a fuzz-only ASAN env with a bounded quarantine (64MB),
  capped redzone (256), small malloc context. Releases freed blocks sooner
  and shrinks per-alloc overhead — matters more on Linux ASAN (larger default
  quarantine) than macOS. NOT applied to gc-stress, where a big quarantine
  widens the cross-process UAF window.
- _FZ_MAX_LEN 1MB -> 128KB: a 1MB mutated JSON builds a huge sw_val tree under
  ASAN. Parser memory-safety bugs reproduce on small inputs (decoders are
  depth-capped at 256), and all seeds are <1KB, so this keeps coverage.
- CI SW_FUZZ_ITERS=5000 (local default stays 20000): 4x less time/exposure on
  the slow 2-core runner; the full seed corpus is always replayed first.
- Fixed a real Linux-only -Wformat warning: fprintf("%llu", pid) where pid is
  uint64_t (== unsigned long on Linux) — cast to unsigned long long. Restores
  the zero-warnings gate on Linux.

Verified: zero-warning build; make fuzz (all 6 targets) no-crash at both 5000
and 20000 iters.

+++ src/swarmrt_lang.c
@@ -2647,7 +2647,7 @@ void sw_val_format(FILE *f, sw_val_t *v) {
     case SW_VAL_FLOAT: fprintf(f, "%.17g", v->v.f); break;
     case SW_VAL_STRING: fprintf(f, "%s", v->v.str); break;
     case SW_VAL_ATOM: fprintf(f, ":%s", v->v.str); break;
-    case SW_VAL_PID: fprintf(f, "<pid:%llu>", v->v.pid ? v->v.pid->pid : 0); break;
+    case SW_VAL_PID: fprintf(f, "<pid:%llu>", (unsigned long long)(v->v.pid ? v->v.pid->pid : 0)); break;
     case SW_VAL_REMOTE_PID:
         fprintf(f, "<rpid:%s:%llu>",
                 v->v.rpid.node ? v->v.rpid.node : "?",
